Finish Django

+++ articles/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_POST, require_GET
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth import get_user_model
 from django.http import HttpResponse
 from .forms import ArticleForm, CommentForm
 from .models import Article, Comment
@@ -23,8 +24,10 @@ def create(request):
         form = ArticleForm(request.POST)
         # embed()
         if form.is_valid():
-            form.save()
-            return redirect('articles:index')
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
+            return redirect('articles:detail', article.pk)
         # else:
         #     context = {'form': form}
         #     return render(request, 'articles/create.html', context)
@@ -48,13 +51,16 @@ def detail(request, article_pk):
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article_pk)
+    if article.user == request.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article_pk)
+        else: # GET method
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:detail', article_pk)
     context = {'form': form}
     return render(request, 'articles/update.html', context)
 
@@ -63,7 +69,11 @@ def update(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if article.user == request.user:
+            article.delete()
+        else:
+            return redirect('articles:detail', article_pk)
+        # article.delete()
     return redirect('articles:index')
 
 
@@ -82,6 +92,7 @@ def commentcreate(request, article_pk):
             # article을 기본 form으로 받지 않고 exclude 했을 경우 article_id 를 따로 저장해줌
             comment = form.save(commit=False)
             comment.article_id = article_pk
+            comment.user = request.user
             comment.save()
     return redirect('articles:detail', article_pk)
 
@@ -91,7 +102,38 @@ def commentcreate(request, article_pk):
 def commentdelete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if comment.user == request.user:
+            comment.delete()
         return redirect('articles:detail', article_pk)
     else:
         return HttpResponse('You are Unauthorized', status=401)
+
+
+def like(request, article_pk):
+    user = request.user
+    article = get_object_or_404(Article, pk=article_pk)
+
+    if article.liked_users.filter(pk=user.pk).exists(): # 1개의 데이터라도 존재하면 True
+    # if user in article.liked_users.all():
+        user.liked_articles.remove(article)
+    else:
+        user.liked_articles.add(article)
+
+    return redirect('articles:detail', article_pk)
+
+
+@login_required
+def follow(request, article_pk, user_pk):
+    # 로그인한 유저가 게시글 유저를 Follow or Unfollow 한다.
+    # 로그인 유저
+    user = request.user
+    # 게시글 주인
+    person = get_object_or_404(get_user_model(), pk=user_pk)
+
+    if user in person.followers.all():
+        # unfollow
+        person.followers.remove(user) 
+    else:
+        # follow
+        person.followers.add(user)
+    return redirect('articles:detail', article_pk)
